Add super early post-processing and plotting script.
Aggregates data from the csv with one row per repetition, to get average times for a given simulation config.
It then plots each sim config as a series on a figure.

Doesn't save to disk yet  or generate plots for the step time within a simulation

Formatting is not great either